requirements.txt
# FOR PUSHOVER

ring-doorbell>=0.8.0
python-dotenv>=1.0.0
requests>=2.31.0
opencv-python-headless>=4.8.0